[Messenger] Move Transport/InMemoryTransport to Transport/InMemory/InMemoryTransport
