+++ UPGRADE-7.0.md
@@ -1,6 +1,14 @@
 UPGRADE FROM 6.4 to 7.0
 =======================
 
+Messenger
+---------
+
+ * Remove `Symfony\Component\Messenger\Transport\InMemoryTransport` and
+   `Symfony\Component\Messenger\Transport\InMemoryTransportFactory` in favor of
+   `Symfony\Component\Messenger\Transport\InMemory\InMemoryTransport` and
+   `Symfony\Component\Messenger\Transport\InMemory\InMemoryTransportFactory`
+
 Workflow
 --------
