+++ src/Symfony/Bundle/FrameworkBundle/composer.json
@@ -48,7 +48,7 @@
         "symfony/http-client": "^5.4|^6.0",
         "symfony/lock": "^5.4|^6.0",
         "symfony/mailer": "^5.4|^6.0",
-        "symfony/messenger": "^6.2",
+        "symfony/messenger": "^6.3",
         "symfony/mime": "^6.2",
         "symfony/notifier": "^5.4|^6.0",
         "symfony/process": "^5.4|^6.0",
@@ -83,7 +83,7 @@
         "symfony/form": "<5.4",
         "symfony/lock": "<5.4",
         "symfony/mailer": "<5.4",
-        "symfony/messenger": "<6.2",
+        "symfony/messenger": "<6.3",
         "symfony/mime": "<6.2",
         "symfony/property-info": "<5.4",
         "symfony/property-access": "<5.4",
